Modularization works, remove test generator, start webdriver generator
