+++ VaadinatorTooling/VaadinatorGeneratorBase/pom.xml
@@ -34,5 +34,10 @@ the License.
 			<version>4.10</version>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.apache.maven</groupId>
+			<artifactId>maven-plugin-api</artifactId>
+			<version>2.0</version>
+		</dependency>
 	</dependencies>
 </project>
