+++ VaadinatorTooling/VaadinatorGeneratorBase/src/main/java/de/akquinet/engineering/vaadinator/mojo/VaadinatorConfig.java
@@ -20,6 +20,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
+import org.apache.maven.plugin.logging.Log;
+
 import de.akquinet.engineering.vaadinator.model.BeanDescription;
 
 public class VaadinatorConfig {
@@ -44,11 +46,12 @@ public class VaadinatorConfig {
 	private boolean genServletBase;
 	private boolean hasDisplayBeans;
 	private boolean hasServiceBeans;
-
+	private Log log;
+	
 	public VaadinatorConfig(String projectName, String basePckg, List<BeanDescription> beanDescriptions,
 			ArtifactType artifactTypeEn, GenType genTypeEn, File targetFolderBaseStart, File targetFolderSrcStart,
 			File targetFolderResStart, Map<String, Object> commonMap, Set<String> displayProfileNames,
-			boolean genServletBase, boolean hasDisplayBeans, boolean hasServiceBeans) {
+			boolean genServletBase, boolean hasDisplayBeans, boolean hasServiceBeans, Log log) {
 		this.projectName = projectName;
 		this.basePckg = basePckg;
 		this.beanDescriptions = beanDescriptions;
@@ -62,6 +65,7 @@ public class VaadinatorConfig {
 		this.genServletBase = genServletBase;
 		this.hasDisplayBeans = hasDisplayBeans;
 		this.hasServiceBeans = hasServiceBeans;
+		this.log = log;
 	}
 
 	public String getProjectName() {
@@ -166,5 +170,9 @@ public class VaadinatorConfig {
 
 	public void setHasServiceBeans(boolean hasServiceBeans) {
 		this.hasServiceBeans = hasServiceBeans;
+	}
+	
+	public Log getLog() {
+		return log;
 	}
 }
