--- VaadinatorTooling/VaadinatorGeneratorTestedit/src/main/resources/META-INF/services/de.akquinet.engineering.vaadinator.generator.CodeGenerator
@@ -1,1 +1,0 @@
-org.vergien.TestCodeGenerator
