+++ VaadinatorTooling/VaadinatorGeneratorWebDriver/src/main/resources/META-INF/services/de.akquinet.engineering.vaadinator.generator.CodeGenerator
@@ -1,0 +1,1 @@
+org.vergien.vaadinator.codegenerator.webdriver.WebDriverCodeGenerator
